boot: bootutil: Fix serial recovery issues
Fixes serial recovery issues in various MCUboot operating modes by
moving functions that are needed out from loader to a common file
and adapting them to work in these different modes
