+++ boot/bootutil/src/bootutil_priv.h
@@ -88,6 +88,12 @@ struct flash_area;
 #define BOOT_STATUS_OP_SWAP     2
 #endif
 
+#if (BOOT_IMAGE_NUMBER > 1)
+#define IMAGES_ITER(x) for ((x) = 0; (x) < BOOT_IMAGE_NUMBER; ++(x))
+#else
+#define IMAGES_ITER(x)
+#endif
+
 /*
  * Maintain state of copy progress.
  */
