+++ boot/zephyr/single_loader.c
@@ -18,15 +18,23 @@
 
 BOOT_LOG_MODULE_DECLARE(mcuboot);
 
-/* Variables passed outside of unit via poiters. */
-static const struct flash_area *_fa_p;
+/* Variables passed outside of unit via pointers. */
 static struct image_header _hdr = { 0 };
-static struct boot_loader_state boot_data;
+static struct boot_loader_state state;
 
 struct boot_loader_state *boot_get_loader_state(void)
 {
-    return &boot_data;
+    return &state;
 }
+
+#if defined(MCUBOOT_SERIAL_IMG_GRP_SLOT_INFO)
+static struct image_max_size image_max_sizes[BOOT_IMAGE_NUMBER] = {0};
+
+struct image_max_size *boot_get_image_max_sizes(void)
+{
+    return image_max_sizes;
+}
+#endif
 
 #if defined(MCUBOOT_VALIDATE_PRIMARY_SLOT) || defined(MCUBOOT_VALIDATE_PRIMARY_SLOT_ONCE)
 /**
@@ -103,6 +111,69 @@ boot_image_validate_once(const struct flash_area *fa_p,
 }
 #endif
 
+int
+boot_open_all_flash_areas(struct boot_loader_state *state)
+{
+    int rc;
+
+    rc = flash_area_open(FLASH_AREA_IMAGE_PRIMARY(0), &BOOT_IMG_AREA(state, BOOT_PRIMARY_SLOT));
+    assert(rc == 0);
+
+    return rc;
+}
+
+void
+boot_close_all_flash_areas(struct boot_loader_state *state)
+{
+    flash_area_close(BOOT_IMG_AREA(state, BOOT_PRIMARY_SLOT));
+}
+
+#if defined(MCUBOOT_SERIAL_IMG_GRP_SLOT_INFO)
+static int app_max_sectors(struct boot_loader_state *state)
+{
+    uint32_t sz = 0;
+    uint32_t sector_sz;
+    uint32_t trailer_sz;
+    uint32_t trailer_sectors = 0;
+
+    sector_sz = boot_img_sector_size(state, BOOT_PRIMARY_SLOT, 0);
+    trailer_sz = boot_trailer_sz(BOOT_WRITE_SZ(state));
+
+    while (1) {
+        sz += sector_sz;
+        ++trailer_sectors;
+
+        if  (sz >= trailer_sz) {
+            break;
+        }
+    }
+
+    return boot_img_num_sectors(state, BOOT_PRIMARY_SLOT) - trailer_sectors;
+}
+
+int app_max_size(struct boot_loader_state *state)
+{
+    uint32_t sector_sz;
+
+    sector_sz = boot_img_sector_size(state, BOOT_PRIMARY_SLOT, 0);
+
+    return app_max_sectors(state) * sector_sz;
+}
+
+/**
+ * Fetches the maximum allowed size of the image
+ */
+const struct image_max_size *boot_get_max_app_size(void)
+{
+    if (image_max_sizes[0].calculated == false) {
+        /* Information not available, need to fetch it */
+        boot_fetch_slot_state_sizes();
+    }
+
+    return image_max_sizes;
+}
+#endif
+
 /**
  * Gather information on image and prepare for booting.
  *
@@ -118,16 +189,13 @@ boot_go(struct boot_rsp *rsp)
 
     BOOT_LOG_DBG("boot_go: Single loader");
 
-    rc = flash_area_open(FLASH_AREA_IMAGE_PRIMARY(0), &_fa_p);
-    assert(rc == 0);
+    rc = boot_open_all_flash_areas(&state);
 
-    rc = boot_image_load_header(_fa_p, &_hdr);
+    rc = boot_image_load_header(BOOT_IMG_AREA(&state, BOOT_PRIMARY_SLOT), &_hdr);
     if (rc != 0)
         goto out;
 
 #ifdef MCUBOOT_RAM_LOAD
-        static struct boot_loader_state state;
-        BOOT_IMG_AREA(&state, 0) = _fa_p;
         state.imgs[0][0].hdr = _hdr;
 
         rc = boot_load_image_to_sram(&state);
@@ -136,7 +204,7 @@ boot_go(struct boot_rsp *rsp)
 #endif
 
 #ifdef MCUBOOT_VALIDATE_PRIMARY_SLOT
-    FIH_CALL(boot_image_validate, fih_rc, _fa_p, &_hdr);
+    FIH_CALL(boot_image_validate, fih_rc, BOOT_IMG_AREA(&state, BOOT_PRIMARY_SLOT), &_hdr);
     if (FIH_NOT_EQ(fih_rc, FIH_SUCCESS)) {
 #ifdef MCUBOOT_RAM_LOAD
         boot_remove_image_from_sram(&state);
@@ -144,7 +212,7 @@ boot_go(struct boot_rsp *rsp)
         goto out;
     }
 #elif defined(MCUBOOT_VALIDATE_PRIMARY_SLOT_ONCE)
-    FIH_CALL(boot_image_validate_once, fih_rc, _fa_p, &_hdr);
+    FIH_CALL(boot_image_validate_once, fih_rc, BOOT_IMG_AREA(&state, BOOT_PRIMARY_SLOT), &_hdr);
     if (FIH_NOT_EQ(fih_rc, FIH_SUCCESS)) {
 #ifdef MCUBOOT_RAM_LOAD
         boot_remove_image_from_sram(&state);
@@ -156,7 +224,7 @@ boot_go(struct boot_rsp *rsp)
 #endif /* MCUBOOT_VALIDATE_PRIMARY_SLOT */
 
 #ifdef MCUBOOT_MEASURED_BOOT
-    rc = boot_save_boot_status(0, &_hdr, _fa_p);
+    rc = boot_save_boot_status(0, &_hdr, BOOT_IMG_AREA(&state, BOOT_PRIMARY_SLOT));
     if (rc != 0) {
         BOOT_LOG_ERR("Failed to add image data to shared area");
         return rc;
@@ -164,19 +232,19 @@ boot_go(struct boot_rsp *rsp)
 #endif /* MCUBOOT_MEASURED_BOOT */
 
 #ifdef MCUBOOT_DATA_SHARING
-    rc = boot_save_shared_data(&_hdr, _fa_p, 0, NULL);
+    rc = boot_save_shared_data(&_hdr, BOOT_IMG_AREA(&state, BOOT_PRIMARY_SLOT), 0, NULL);
     if (rc != 0) {
         BOOT_LOG_ERR("Failed to add data to shared memory area.");
         return rc;
     }
 #endif /* MCUBOOT_DATA_SHARING */
 
-    rsp->br_flash_dev_id = flash_area_get_device_id(_fa_p);
-    rsp->br_image_off = flash_area_get_off(_fa_p);
+    rsp->br_flash_dev_id = flash_area_get_device_id(BOOT_IMG_AREA(&state, BOOT_PRIMARY_SLOT));
+    rsp->br_image_off = flash_area_get_off(BOOT_IMG_AREA(&state, BOOT_PRIMARY_SLOT));
     rsp->br_hdr = &_hdr;
 
 out:
-    flash_area_close(_fa_p);
+    boot_close_all_flash_areas(&state);
 
     FIH_RET(fih_rc);
 }
